Task7 fix integer temp

+++ task7/src/main/java/com/hplasplas/task7/models/weather/forecast/Main.java
@@ -29,7 +29,7 @@ public class Main {
     private long humidity;
     @SerializedName("temp_kf")
     @Expose
-    private Integer tempKf;
+    private double tempKf;
 
     public double getTemp() {
         return temp;
@@ -87,11 +87,11 @@ public class Main {
         this.humidity = humidity;
     }
 
-    public Integer getTempKf() {
+    public double getTempKf() {
         return tempKf;
     }
 
-    public void setTempKf(Integer tempKf) {
+    public void setTempKf(double tempKf) {
         this.tempKf = tempKf;
     }
